[Coupling] Modified the process definitions in Process/HT for the assembly for both of the monolithic scheme and the staggered scheme.
