+++ ProcessLib/HT/StaggeredHTFEM.h
@@ -1,0 +1,87 @@
+/**
+ * \copyright
+ * Copyright (c) 2012-2017, OpenGeoSys Community (http://www.opengeosys.org)
+ *            Distributed under a Modified BSD License.
+ *              See accompanying file LICENSE.txt or
+ *              http://www.opengeosys.org/project/license
+ *
+ *  \file   StaggeredHTFEM.h
+ *  Created on October 11, 2017, 1:42 PM
+ */
+
+#pragma once
+
+#include <Eigen/Dense>
+#include <vector>
+
+#include "HTProcessData.h"
+#include "NumLib/DOF/DOFTableUtil.h"
+#include "NumLib/Extrapolation/ExtrapolatableElement.h"
+#include "NumLib/Fem/FiniteElement/TemplateIsoparametric.h"
+#include "NumLib/Fem/ShapeMatrixPolicy.h"
+#include "ProcessLib/Parameter/Parameter.h"
+#include "ProcessLib/Utils/InitShapeMatrices.h"
+
+#include "HTFEM.h"
+
+namespace ProcessLib
+{
+struct LocalCoupledSolutions;
+
+namespace HT
+{
+template <typename ShapeFunction, typename IntegrationMethod,
+          unsigned GlobalDim>
+class StaggeredHTFEM : public HTFEM<ShapeFunction, IntegrationMethod, GlobalDim>
+{
+    using ShapeMatricesType = ShapeMatrixPolicyType<ShapeFunction, GlobalDim>;
+    using ShapeMatrices = typename ShapeMatricesType::ShapeMatrices;
+
+    using LocalMatrixType =
+        typename ShapeMatricesType::template MatrixType<ShapeFunction::NPOINTS,
+                                                        ShapeFunction::NPOINTS>;
+    using LocalVectorType =
+        typename ShapeMatricesType::template VectorType<ShapeFunction::NPOINTS>;
+
+    using NodalVectorType = typename ShapeMatricesType::NodalVectorType;
+    using NodalRowVectorType = typename ShapeMatricesType::NodalRowVectorType;
+
+    using GlobalDimVectorType = typename ShapeMatricesType::GlobalDimVectorType;
+    using GlobalDimNodalMatrixType =
+        typename ShapeMatricesType::GlobalDimNodalMatrixType;
+    using GlobalDimMatrixType = typename ShapeMatricesType::GlobalDimMatrixType;
+
+public:
+    StaggeredHTFEM(MeshLib::Element const& element,
+                   std::size_t const local_matrix_size,
+                   bool is_axially_symmetric,
+                   unsigned const integration_order,
+                   HTMaterialProperties const& process_data)
+        : HTFEM<ShapeFunction, IntegrationMethod, GlobalDim>(
+              element, local_matrix_size, is_axially_symmetric,
+              integration_order, process_data, 1)
+    {
+    }
+
+    void assembleWithCoupledTerm(
+        double const t, std::vector<double>& local_M_data,
+        std::vector<double>& local_K_data, std::vector<double>& local_b_data,
+        LocalCoupledSolutions const& coupled_term) override;
+
+private:
+    void assembleHydraulicEquation(double const t,
+                                   std::vector<double>& local_M_data,
+                                   std::vector<double>& local_K_data,
+                                   std::vector<double>& local_b_data,
+                                   LocalCoupledSolutions const& coupled_term);
+
+    void assembleHeatTransportEquation(
+        double const t, std::vector<double>& local_M_data,
+        std::vector<double>& local_K_data, std::vector<double>& local_b_data,
+        LocalCoupledSolutions const& coupled_term);
+};
+
+}  // namespace HT
+}  // namespace ProcessLib
+
+#include "StaggeredHTFEM-impl.h"
